refactor: extract into subcommand
